requirements.txt
Flask==2.3.3
boto3==1.28.57
python-dotenv==1.0.0
gunicorn==21.2.0
